Create Proposition.py
Making a proposition.
Use the comparison operator we learned earlier to create a proposition that satisfies the given conditions and put it in! A proposition is a sentence or expression that can determine whether the content is true or false.

Let's put a proposition that satisfies the conditions given to the three variables on the right!